add CommandHandler and Remove_by_id
