+++ src/main/java/commands/Remove_by_id.java
@@ -1,6 +1,7 @@
 package commands;
 
 import dataBase.DataBase;
+import defaultClasses.Person;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
@@ -11,12 +12,16 @@ public class Remove_by_id extends BaseCommand {
 
     public void execute(DataBase obj) throws IOException {
         try{
-            System.out.println(super.getParameter());
-            Integer a = Integer.parseInt(super.getParameter());
-            var temp = obj.getCollection();
-            Iterator i = temp.iterator();
-            Field field = i.next().getClass().getDeclaredField("id");
-            System.out.println(field);
+            Long a = Long.parseLong(super.getParameter());
+            var array = obj.getCollection();
+            Iterator<Person> it = array.iterator();
+            while(it.hasNext()){
+                if(it.next().getId().equals(a)){
+                    it.remove();
+                    break;
+                }
+            }
+            System.out.println("Element with id " + a + " removed");
         }catch (Exception e){
             System.out.println("Incorrect argument, command cannot be executed");
         }
