+++ src/main/java/dataBase/GlobalObj.java
@@ -3,7 +3,5 @@ package dataBase;
 import defaultClasses.Person;
 
 public class GlobalObj {
-    public static DataBase<Person> dataBase = new DataBase<>();
-
-
+    public static DataBase dataBase = new DataBase();
 }
